fix delete

+++ src/components/StoreItem.vue
@@ -12,11 +12,7 @@
           class="ml-2 mr-2"
         ></v-rating>
         <v-col cols="1"></v-col>
-        <v-icon
-          icon="mdi-cash"
-          color="green-lighten-3"
-          class="mr-2"
-        ></v-icon>
+        <v-icon icon="mdi-cash" color="green-lighten-3" class="mr-2"></v-icon>
         <span class="mr-2">{{ product.data.price }}</span>
         <v-col cols="1"></v-col>
         <v-icon
@@ -27,7 +23,7 @@
         <span>{{ product.data.stock }}</span>
       </v-row>
       <v-col cols="1"></v-col>
-      
+
       <v-img
         :src="product.data.image"
         width="200"
@@ -41,18 +37,22 @@
 </template>
 
 <script lang="ts" setup>
-import { ProductDoc } from '../types/product';
-import { useProductStore } from '../stores/ProductStore';
+import { ProductDoc } from "../types/product";
+import { useProductStore } from "../stores/ProductStore";
 
 const { product } = defineProps<{
-  product: ProductDoc
+  product: ProductDoc;
 }>();
 
 const productStore = useProductStore();
 
 // Method to delete the current product
-const deleteProduct = async () =>{
-  console.log('Deleting product:', product.id);
-  await productStore.deleteProduct(product.id);
+const deleteProduct = async () => {
+  // console.log('Deleting product:', product.id);
+  // await productStore.deleteProduct(product.id);
+
+  if (confirm("Do you want to delete this item?" + product.id)) {
+    productStore.deleteProduct(product.id);
+  }
 };
 </script>
